fix: ugly hack to fix `Resources` list.
The issue: we correctly iterate over `.Pages`, but `.Paginator.Pages` isn't
set correctly. (recognizes `yee.md` files, but not `yee/_index.md` dirs for
some reason).

I re-initialize the Hugo paginator in this directory, so it'll scan the subdirs
and pick up the index dirs. Now, it'll go over everything in the list.
Hacky fix...